requirements.txt
streamlit==1.50.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.3.0

# HuggingFace Transformers + PyTorch for Emotion Model
transformers>=4.35.0
torch>=2.0.0

# Optional but recommended (used internally by Transformers)
huggingface-hub>=0.17.0
tqdm>=4.65.0
